pipeline/requirements.txt
requests>=2.28.0
numpy>=1.21.0
python-dotenv>=1.0.0
streamlit>=1.20.0
pandas>=1.5.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
datasets>=2.14.0
